src/server/filetransfer/requirements.txt
async-generator==1.10
attrs==22.1.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
et-xmlfile==1.1.0
Flask==2.2.3
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.23.5
openpyxl==3.0.10
outcome==1.2.0
packaging==22.0
pandas==1.5.2
pdfkit==1.0.0
PyPDF2==2.12.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.13
webdriver-manager==3.8.5
Werkzeug==2.2.3
wsproto==1.2.0
